provide better error handling
