+++ mm2src/coins/lp_coins.rs
@@ -241,15 +241,15 @@ pub use solana::{solana_coin_from_conf_and_params, SolanaActivationParams, Solan
 pub mod utxo;
 #[cfg(not(target_arch = "wasm32"))] pub mod z_coin;
 
-use eth::{eth_coin_from_conf_and_request, eth_coin_from_conf_and_request_v2, EnableV2RpcRequest, EthCoin,
-          EthTxFeeDetails, SignedEthTx};
+use eth::{eth_coin_from_conf_and_request, eth_coin_from_conf_and_request_v2, EthCoin, EthTxFeeDetails, SignedEthTx};
 use hd_wallet::{HDAddress, HDAddressId};
 use qrc20::Qrc20ActivationParams;
 use qrc20::{qrc20_coin_from_conf_and_params, Qrc20Coin, Qrc20FeeDetails};
 use qtum::{Qrc20AddressError, ScriptHashTypeNotSupported};
-use rpc_command::init_create_account::{CreateAccountTaskManager, CreateAccountTaskManagerShared};
-use rpc_command::init_scan_for_new_addresses::{ScanAddressesTaskManager, ScanAddressesTaskManagerShared};
-use rpc_command::init_withdraw::{WithdrawTaskManager, WithdrawTaskManagerShared};
+use rpc_command::{enable_v2::EnableV2RpcRequest,
+                  init_create_account::{CreateAccountTaskManager, CreateAccountTaskManagerShared},
+                  init_scan_for_new_addresses::{ScanAddressesTaskManager, ScanAddressesTaskManagerShared},
+                  init_withdraw::{WithdrawTaskManager, WithdrawTaskManagerShared}};
 use utxo::bch::{bch_coin_from_conf_and_params, BchActivationRequest, BchCoin};
 use utxo::qtum::{self, qtum_coin_with_priv_key, QtumCoin};
 use utxo::qtum::{QtumDelegationOps, QtumDelegationRequest, QtumStakingInfosDetails};
