+++ mm2src/coins/rpc_command/mod.rs
@@ -1,4 +1,5 @@
 pub mod account_balance;
+pub mod enable_v2;
 pub mod hd_account_balance_rpc_error;
 pub mod init_create_account;
 pub mod init_scan_for_new_addresses;
